Created the screen_trading.xml layout. Implemented NumberPickers from the merchantInteraction.java class.
Added Collection to Merchant.java so that all merchants are created with first/last name, location, and a randomized purse size of 500-750 gold.

Also added protected File newFile to the createNewGame.java class so that after the player hits the complete button, they create a new file with their playername as the path. If playerName is empty, it assigns name as "Haven".

Implemented a WarningDialog on the MainActivity.java for when the player wants to start a new game.

Pending final design of screen_trading.xml and merchant images. Also pending decision of where to instantiate the merchant collection & how to use that to differentiate between cities w 1 merchant those w 2 merchants.